Use size_t for sizes.

+++ cpukit/posix/src/mqueuesendsupp.c
@@ -41,7 +41,7 @@
 int _POSIX_Message_queue_Send_support(
   mqd_t               mqdes,
   const char         *msg_ptr,
-  uint32_t            msg_len,
+  size_t              msg_len,
   uint32_t            msg_prio,
   Watchdog_Interval   timeout
 )
